// src/enums/amino-acid-side-chain-type.ts
/**
 * Enum representing the chemical classification of amino acid side chains
 */
export enum AminoAcidSideChainType {
    ALIPHATIC = 'aliphatic',
    AROMATIC = 'aromatic',
    BASIC = 'basic',
    ACIDIC = 'acidic',
    AMIDE = 'amide',
    SULFUR_CONTAINING = 'sulfur-containing',
    HYDROXYL_CONTAINING = 'hydroxyl-containing',
    IMINO = 'imino'
}